// refactor/06_settlement_audit/enums.ts
export enum HistoricalSampleRejectionReason {
  NOT_FORMAL_RECOMMENDATION = 'NOT_FORMAL_RECOMMENDATION',
  SCORE_NOT_VERIFIED = 'SCORE_NOT_VERIFIED',
  SCORE_INVALID = 'SCORE_INVALID',
  SETTLEMENT_NOT_RESOLVED = 'SETTLEMENT_NOT_RESOLVED',
  SETTLEMENT_NOT_BINARY = 'SETTLEMENT_NOT_BINARY',
  PREDICTION_TIMESTAMP_INVALID = 'PREDICTION_TIMESTAMP_INVALID',
  LIVE_MINUTE_INVALID = 'LIVE_MINUTE_INVALID',
  MODEL_PROBABILITY_INVALID = 'MODEL_PROBABILITY_INVALID',
  PREDICTED_LAMBDA_INVALID = 'PREDICTED_LAMBDA_INVALID'
}